+++ sdk/core/core-http/src/policies/proxyPolicy.ts
@@ -14,8 +14,7 @@ import { Constants } from "../util/constants";
 import { URLBuilder } from "../url";
 import { getEnvironmentValue } from "../util/utils";
 
-let noProxyList: string[] = [];
-let isNoProxyInitalized = false;
+export const noProxyList: string[] = loadNoProxy();
 const byPassedList: Map<string, boolean> = new Map();
 
 function loadEnvironmentProxyValue(): string | undefined {
@@ -30,14 +29,17 @@ function loadEnvironmentProxyValue(): string | undefined {
   return httpsProxy || allProxy || httpProxy;
 }
 
-// Check whether the given `uri` matches the noProxyList. If it matches, any request sent to that same `uri` won't set the proxy settings.
+// Check whether the host of a given `uri` is in the noProxyList.
+// If there's a match, any request sent to the same host won't have the proxy settings set.
 function isBypassed(uri: string): boolean | undefined {
-  if (byPassedList.has(uri)) {
-    return byPassedList.get(uri);
+  if (noProxyList.length === 0) {
+    return false;
   }
-  loadNoProxy();
-  let isBypassedFlag = false;
   const host = URLBuilder.parse(uri).getHost()!;
+  if (byPassedList.has(host)) {
+    return byPassedList.get(host);
+  }
+  let isBypassedFlag = false;
   for (const proxyString of noProxyList) {
     if (proxyString[0] === ".") {
       if (uri.endsWith(proxyString)) {
@@ -53,20 +55,20 @@ function isBypassed(uri: string): boolean | undefined {
       }
     }
   }
-  byPassedList.set(uri, isBypassedFlag);
+  byPassedList.set(host, isBypassedFlag);
   return isBypassedFlag;
 }
 
-function loadNoProxy(): void {
-  if (isNoProxyInitalized) {
-    return;
-  }
+export function loadNoProxy(): string[] {
   const noProxy = getEnvironmentValue(Constants.NO_PROXY);
   if (noProxy) {
-    const list = noProxy.split(",");
-    noProxyList = list.map((item) => item.trim()).filter((item) => item.length);
+    return noProxy
+      .split(",")
+      .map((item) => item.trim())
+      .filter((item) => item.length);
   }
-  isNoProxyInitalized = true;
+
+  return [];
 }
 
 export function getDefaultProxySettings(proxyUrl?: string): ProxySettings | undefined {