+++ sdk/core/core-https/src/policies/proxyPolicy.ts
@@ -7,28 +7,78 @@ import { URL } from "../util/url";
 
 const HTTPS_PROXY = "HTTPS_PROXY";
 const HTTP_PROXY = "HTTP_PROXY";
+const ALL_PROXY = "ALL_PROXY";
+const NO_PROXY = "NO_PROXY";
 
 /**
  * The programmatic identifier of the proxyPolicy.
  */
 export const proxyPolicyName = "proxyPolicy";
 
+export const noProxyList: string[] = loadNoProxy();
+const byPassedList: Map<string, boolean> = new Map();
+
+function getEnvironmentValue(name: string): string | undefined {
+  if (process.env[name]) {
+    return process.env[name];
+  } else if (process.env[name.toLowerCase()]) {
+    return process.env[name.toLowerCase()];
+  }
+  return undefined;
+}
+
 function loadEnvironmentProxyValue(): string | undefined {
   if (!process) {
     return undefined;
   }
 
-  if (process.env[HTTPS_PROXY]) {
-    return process.env[HTTPS_PROXY];
-  } else if (process.env[HTTPS_PROXY.toLowerCase()]) {
-    return process.env[HTTPS_PROXY.toLowerCase()];
-  } else if (process.env[HTTP_PROXY]) {
-    return process.env[HTTP_PROXY];
-  } else if (process.env[HTTP_PROXY.toLowerCase()]) {
-    return process.env[HTTP_PROXY.toLowerCase()];
+  const httpsProxy = getEnvironmentValue(HTTPS_PROXY);
+  const allProxy = getEnvironmentValue(ALL_PROXY);
+  const httpProxy = getEnvironmentValue(HTTP_PROXY);
+
+  return httpsProxy || allProxy || httpProxy;
+}
+
+// Check whether the host of a given `uri` is in the noProxyList.
+// If there's a match, any request sent to the same host won't have the proxy settings set.
+function isBypassed(uri: string): boolean | undefined {
+  if (noProxyList.length === 0) {
+    return false;
+  }
+  const host = new URL(uri).hostname;
+  if (byPassedList.has(host)) {
+    return byPassedList.get(host);
+  }
+  let isBypassedFlag = false;
+  for (const proxyString of noProxyList) {
+    if (proxyString[0] === ".") {
+      if (uri.endsWith(proxyString)) {
+        isBypassedFlag = true;
+      } else {
+        if (host === proxyString.slice(1) && host.length === proxyString.length - 1) {
+          isBypassedFlag = true;
+        }
+      }
+    } else {
+      if (host === proxyString) {
+        isBypassedFlag = true;
+      }
+    }
+  }
+  byPassedList.set(host, isBypassedFlag);
+  return isBypassedFlag;
+}
+
+export function loadNoProxy(): string[] {
+  const noProxy = getEnvironmentValue(NO_PROXY);
+  if (noProxy) {
+    return noProxy
+      .split(",")
+      .map((item) => item.trim())
+      .filter((item) => item.length);
   }
 
-  return undefined;
+  return [];
 }
 
 /**
@@ -65,7 +115,7 @@ export function proxyPolicy(proxySettings = getDefaultProxySettings()): Pipeline
   return {
     name: proxyPolicyName,
     async sendRequest(request: PipelineRequest, next: SendRequest): Promise<PipelineResponse> {
-      if (!request.proxySettings) {
+      if (!request.proxySettings && !isBypassed(request.url)) {
         request.proxySettings = proxySettings;
       }
       return next(request);