+++ sdk/core/core-client/src/interfaces.ts
@@ -50,10 +50,9 @@ export type RequiredSerializerOptions = {
 };
 
 /**
- * This interface extends a generic `PipelineRequest` to include
- * additional metadata about the request.
+ * A type alias for future proofing.
  */
-export type OperationRequest = PipelineRequest<OperationRequestInfo>;
+export type OperationRequest = PipelineRequest;
 
 /**
  * Metadata that is used to properly parse a response.
@@ -105,6 +104,13 @@ export interface OperationOptions {
    * Options to override serialization/de-serialization behavior.
    */
   serializerOptions?: SerializerOptions;
+
+  /**
+   * A function to be called each time a response is received from the server
+   * while performing the requested operation.
+   * May be called multiple times.
+   */
+  onResponse?: RawResponseCallback;
 }
 
 /**
@@ -333,17 +339,14 @@ export interface FullOperationResponse extends PipelineResponse {
 }
 
 /**
- * The processed and flattened response to an operation call.
- * Contains merged properties of the parsed body and headers.
+ * A function to be called each time a response is received from the server
+ * while performing the requested operation.
+ * May be called multiple times.
  */
-export interface OperationResponse {
-  /**
-   * The underlying HTTP response containing both raw and deserialized response data.
-   */
-  _response: FullOperationResponse;
-
-  [key: string]: any;
-}
+export type RawResponseCallback = (
+  rawResponse: FullOperationResponse,
+  flatResponse: unknown
+) => void;
 
 /**
  * Used to map raw response objects to final shapes.