+++ sdk/monitor/opentelemetry-exporter-azure-monitor/src/export/trace.ts
@@ -2,8 +2,9 @@
 // Licensed under the MIT license.
 
 import { Logger } from "@opentelemetry/api";
-import { ConsoleLogger, LogLevel, ExportResult } from "@opentelemetry/core";
+import { ConsoleLogger, LogLevel, ExportResult, ExportResultCode } from "@opentelemetry/core";
 import { ReadableSpan, SpanExporter } from "@opentelemetry/tracing";
+import { RestError } from "@azure/core-http";
 import { ConnectionStringParser } from "../utils/connectionStringParser";
 import { HttpSender, FileSystemPersist } from "../platform";
 import {
@@ -64,9 +65,14 @@ export class AzureMonitorTraceExporter implements SpanExporter {
   private async _persist(envelopes: unknown[]): Promise<ExportResult> {
     try {
       const success = await this._persister.push(envelopes);
-      return success ? ExportResult.FAILED_RETRYABLE : ExportResult.FAILED_NOT_RETRYABLE;
-    } catch (e) {
-      return ExportResult.FAILED_NOT_RETRYABLE;
+      return success
+        ? { code: ExportResultCode.SUCCESS }
+        : {
+            code: ExportResultCode.FAILED,
+            error: new Error("Failed to persist envelope in disk.")
+          };
+    } catch (ex) {
+      return { code: ExportResultCode.FAILED, error: ex };
     }
   }
 
@@ -84,7 +90,7 @@ export class AzureMonitorTraceExporter implements SpanExporter {
           }, this._options.batchSendRetryIntervalMs);
           this._retryTimer.unref();
         }
-        return ExportResult.SUCCESS;
+        return { code: ExportResultCode.SUCCESS };
       } else if (statusCode && isRetriable(statusCode)) {
         // Failed -- persist failed data
         if (result) {
@@ -102,15 +108,24 @@ export class AzureMonitorTraceExporter implements SpanExporter {
         }
       } else {
         // Failed -- not retriable
-        return ExportResult.FAILED_NOT_RETRYABLE;
+        return {
+          code: ExportResultCode.FAILED
+        };
       }
     } catch (senderErr) {
-      // Request failed -- always retry
-      this._logger.error(
-        "Envelopes could not be exported and are not retriable. Error message:",
-        senderErr.message
-      );
-      return ExportResult.FAILED_NOT_RETRYABLE;
+      if (this._isNetworkError(senderErr)) {
+        this._logger.error(
+          "Retrying due to transient client side error. Error message:",
+          senderErr.message
+        );
+        return await this._persist(envelopes);
+      } else {
+        this._logger.error(
+          "Envelopes could not be exported and are not retriable. Error message:",
+          senderErr.message
+        );
+        return { code: ExportResultCode.FAILED, error: senderErr };
+      }
     }
   }
 
@@ -125,9 +140,9 @@ export class AzureMonitorTraceExporter implements SpanExporter {
     resultCallback(await this.exportEnvelopes(envelopes));
   }
 
-  shutdown(): void {
+  async shutdown(): Promise<void> {
     this._logger.info("Azure Monitor Trace Exporter shutting down");
-    this._sender.shutdown();
+    return this._sender.shutdown();
   }
 
   private async _sendFirstPersistedFile(): Promise<void> {
@@ -139,5 +154,14 @@ export class AzureMonitorTraceExporter implements SpanExporter {
     } catch (err) {
       this._logger.warn(`Failed to fetch persisted file`, err);
     }
+  }
+
+  private _isNetworkError(error: Error): boolean {
+    if (error instanceof RestError) {
+      if (error && error.code && error.code === "REQUEST_SEND_ERROR") {
+        return true;
+      }
+    }
+    return false;
   }
 }