+++ sdk/servicebus/service-bus/src/util/atomXmlHelper.ts
@@ -19,6 +19,7 @@ import { administrationLogger as logger } from "../log";
 import { Buffer } from "buffer";
 
 import { parseURL } from "./parseUrl";
+import { isJSONLikeObject } from "./utils";
 
 /**
  * @internal
@@ -94,6 +95,28 @@ export async function executeAtomXmlOperation(
 /**
  * @internal
  * @hidden
+ * The key-value pairs having undefined/null as the values would lead to the empty tags in the serialized XML request.
+ * Empty tags in the request body is problematic because of the following reasons.
+ * - ATOM based management operations throw a "Bad Request" error if empty tags are included in the XML request body at top level.
+ * - At the inner levels, Service assigns the empty strings as values to the empty tags instead of throwing an error.
+ *
+ * This method recursively removes the key-value pairs with undefined/null as the values from the request object that is to be serialized.
+ *
+ * @param {{ [key: string]: any }} resource
+ */
+export function sanitizeSerializableObject(resource: { [key: string]: any }) {
+  Object.keys(resource).forEach(function(property) {
+    if (resource[property] == undefined) {
+      delete resource[property];
+    } else if (isJSONLikeObject(resource[property])) {
+      sanitizeSerializableObject(resource[property]);
+    }
+  });
+}
+
+/**
+ * @internal
+ * @hidden
  * Serializes input information to construct the Atom XML request
  * @param resourceName Name of the resource to be serialized like `QueueDescription`
  * @param resource The entity details
@@ -103,15 +126,8 @@ export async function executeAtomXmlOperation(
 export function serializeToAtomXmlRequest(resourceName: string, resource: any): object {
   const content: any = {};
 
-  // The top level key value pairs having undefined/null as the value are removed in order to address issue where the Service Bus'
-  // ATOM based management operations throw a "Bad Request" error if empty tags are included in the xml request body at top level.
-  const processedResource = Object.assign({}, resource);
-  Object.keys(processedResource).forEach(function(property) {
-    if (processedResource[property] == undefined) {
-      delete processedResource[property];
-    }
-  });
-  content[resourceName] = processedResource;
+  content[resourceName] = Object.assign({}, resource);
+  sanitizeSerializableObject(content[resourceName]);
 
   content[resourceName][Constants.XML_METADATA_MARKER] = {
     xmlns: "http://schemas.microsoft.com/netservices/2010/10/servicebus/connect",