+++ sdk/core/core-http/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 1.2.3 (Unreleased)
 
+- Don't set a default content-type when there is no request body. [PR 13233](https://github.com/Azure/azure-sdk-for-js/pull/13233)
+
+- Clean up abort event handler properly for streaming operations. Fixed [issue 12029](https://github.com/Azure/azure-sdk-for-js/issues/12029)
+
+- Reduce memory usage of the cache in proxy policy. Fixed [issue 13277](https://github.com/Azure/azure-sdk-for-js/issues/13277)
 
 ## 1.2.2 (2021-01-07)